requirements.txt
cachetools==5.5.2
google-api-core==2.24.2
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-adk==0.4.0
google-cloud-storage==2.19.0
google-cloud-texttospeech==2.27.0
google-cloud-video-transcoder==1.15.1
google-generativeai==1.14.0
googleapis-common-protos==1.70.0
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
mutagen==1.47.0
protobuf==6.31.0rc2
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
tinytag==2.1.1
